# PERF: improves performance in SeriesGroupBy.transform